폰트 사이즈 구성 파일 추가 및 기존 설정 변경
